requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dateutil>=2.8.0
google-api-python-client==2.126.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
gdown==4.7.3
extra-streamlit-components
folium>=0.14.0
streamlit-folium>=0.13.0
numpy==1.26.4
requests
pyarrow>=10.0.0
psutil>=5.9.0
duckdb>=0.10.0
gspread
google-cloud-bigquery
playwright
